requirements.txt
fastapi[standard]
python-dotenv
pandas
# python-socketio